Refactored monolith to modular structures for better readability
