+++ config.py
@@ -1,0 +1,26 @@
+import os
+import json
+
+# Default paths and settings
+WHATSAPP_STATUS_PATH = os.path.expandvars(r'%userprofile%\AppData\Local\Packages\5319275A.WhatsAppDesktop_cv1g1gvanyjgm\LocalState\shared\transfers')
+DEFAULT_SAVE_DIR = os.path.join(os.path.expanduser('~'), 'Downloads', 'WhatsappStatuses')
+SETTINGS_DIR = os.path.join(os.path.expanduser('~'), 'AppData', 'Local', 'WhatsAppStatusSaver')
+SETTINGS_FILE = os.path.join(SETTINGS_DIR, "settings.json")
+THUMBNAIL_CACHE_DIR = os.path.join(SETTINGS_DIR, "thumbnail_cache")
+
+def load_settings():
+    if os.path.exists(SETTINGS_FILE):
+        with open(SETTINGS_FILE, 'r') as f:
+            return json.load(f)
+    return {"save_dir": DEFAULT_SAVE_DIR, "theme_mode": "light"}
+
+def save_settings(settings):
+    with open(SETTINGS_FILE, 'w') as f:
+        json.dump(settings, f)
+
+# Create necessary directories
+if not os.path.exists(SETTINGS_DIR):
+    os.makedirs(SETTINGS_DIR)
+
+if not os.path.exists(THUMBNAIL_CACHE_DIR):
+    os.makedirs(THUMBNAIL_CACHE_DIR)
