+++ flet.config
@@ -1,0 +1,5 @@
+[build]
+name = WhatsApp Status Saver
+description = Status downloader for WhatsApp desktop
+version = 0.1.0
+author = Alvin Mayende
