+++ main.py
@@ -1,0 +1,6 @@
+import flet as ft
+import asyncio
+from app import main
+
+if __name__ == "__main__":
+    asyncio.run(ft.app(target=main))
