+++ requirements.txt
@@ -1,0 +1,3 @@
+flet
+Pillow
+opencv-python
